Rename dataAggregation.ts to supabaseUtils.ts
